feat: Enhance TranslationQueueManager with job handling and improve request registration
- Added optional job parameter to registerRequest for tracking translation jobs.
- Improved handling of translation requests to ensure proper cancellation of metadata translations when content translation starts.
- Updated active request management to include job tracking.
- Refactored translation request registration logic for clarity and efficiency.

refactor: Simplify Collection Extension Functions

- Updated replaceAll function to directly return a new list from newItems.
- Improved error handling in next and previous functions to throw specific exceptions.

test: Update SourceHealth and SyncRepository Tests

- Refactored assertions in SourceHealthTest for clarity.
- Enhanced SyncRepositoryTest to assert non-null results from observed flows.

fix: Correct PluginDetailsScreen and ViewModel Logic

- Improved null checks and resource usage handling in PluginDetailsScreen.
- Refactored PluginDetailsViewModel to ensure proper handling of remote plugin information.

chore: Remove Unused LazyList Extensions

- Deleted isScrolledToEnd and related scrolling utility functions to streamline codebase.

style: General Code Cleanup and Formatting

- Applied consistent formatting and removed unnecessary comments across various files.
- Improved readability by restructuring conditional checks and variable assignments.

+++ data/src/commonMain/kotlin/ireader/data/downloads/downloadMapper.kt
@@ -7,9 +7,10 @@ val downloadMapper = {  chapterId: Long, bookId: Long, priority: Int, _id: Long,
     SavedDownloadWithInfo(
         chapterId,
         bookId,
-        priority.toInt(),
+        priority,
         _id,
         sourceId = source,
+
         bookName = title,
         url,
         name,
